version 1.1.2.1 add libbyte_proto_analyzer

+++ src/liballoc/sys_alloc/allocator_sys_alloc.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "liballoc/inc_files.h"
 
-void *sys_alloc_alloc(allocator_t *alloc,uint8_t size)
+void *sys_alloc_alloc(allocator_t *alloc,uint32_t size)
 {
 	return malloc(size);
 }
